Add simulators in their own class with documentation
